jamil prep 2019: study_one and study_two scripts ok

+++ analysis/scr/adapting_fyp_analysis.py
@@ -201,7 +201,7 @@ class experimental(object):
         return p1,m1
 
     def findAmplitude(self,data,stims,iStim,nsForward): 
-        interval4max = data[stims[iStim]+1000*self.windowstart:stims[iStim]+1000*nsForward]
+        interval4max = data[int(stims[iStim]+1000*self.windowstart):int(stims[iStim]+1000*nsForward)]
         index4max = stims[iStim] + interval4max.argmax() + 1000*self.windowstart
         value4max = data[index4max]
